requirements.txt
anyio==3.6.2
asyncpg==0.27.0
black==22.12.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorlog==6.7.0
distlib==0.3.6
fastapi==0.88.0
filelock==3.8.2
flake8==6.0.0
h11==0.14.0
httptools==0.5.0
identify==2.5.11
idna==3.4
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.3
platformdirs==2.6.0
pre-commit==2.21.0
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
watchfiles==0.18.1
websockets==10.4
